Move listAll to SDK
